requirements.txt
absl-py==0.7.1
advertorch==0.1.5
alabaster==0.7.12
alembic==1.1.0
argh==0.26.2
asn1crypto==1.3.0
astor==0.8.0
astroid==2.3.3
astropy==4.0
atomicwrites==1.3.0
attrs==19.3.0
autopep8==1.4.4
Babel==2.8.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
bayesian-optimization==1.2.0
beautifulsoup4==4.8.2
bidict==0.18.0
bitarray==1.2.0
bkcharts==0.2
bleach==3.1.0
bokeh==1.4.0
boto==2.49.0
Bottleneck==1.3.1
cached-property==1.5.1
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
chainer==7.0.0
chainercv==0.13.1
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
clyent
cma==3.0.3
colorama==0.4.3
coloredlogs==14.0
configparser
contextlib2==0.6.0.post1
cox==0.1.post3
cryptography==2.8
cupy-cuda111==8.3.0
cvxopt==1.2.5
cvxpy==1.1.7
cycler==0.10.0
Cython==0.29.14
cytoolz==0.10.1
dask==2.9.1
data==0.4
databricks-cli==0.9.0
decorator==4.4.1
defusedxml==0.6.0
diff-match-patch==20181111
dill==0.3.3
distributed==2.9.1
docker==4.0.2
docopt==0.6.2
docutils==0.15.2
eagerpy==0.29.0
easydict==1.9
ecos==2.0.7.post1
entrypoints==0.3
et-xmlfile==1.0.1
faiss-gpu==1.6.3
fastcache==1.1.0
fastrlock==0.5
fcn==6.4.17
filelock==3.0.12
flake8==3.7.9
Flask==1.1.1
foolbox==3.2.1
frozendict==1.2
fsspec==0.6.2
funcsigs==1.0.2
future==0.18.2
gast==0.2.2
gdown==3.8.3
gevent==1.4.0
gitdb==4.0.2
gitdb2==2.0.5
GitPython==3.1.0
glob2==0.7
glog==0.3.1
gmpy2==2.0.8
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.7
gorilla==0.3.0
GPUtil==1.4.0
gpytorch==1.3.0
greenlet==0.4.15
grpcio==1.26.0
gunicorn==19.9.0
gym==0.17.1
h5py==2.9.0
HeapDict==1.0.1
html5lib==1.0.1
humanfriendly==8.1
hypothesis==4.54.2
idna==2.8
imageio==2.6.1
imagesize==1.2.0
img2pdf==0.3.3
importlib-metadata==1.3.0
intervaltree==3.0.2
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.14.1
jeepney==0.4.2
Jinja2==2.10.3
joblib==0.14.1
json5==0.8.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
jupyterlab==1.2.4
jupyterlab-server==1.0.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==20.0.0
kiwisolver==1.1.0
kornia==0.1.4.post2
latex==0.7.0
lazy-object-proxy==1.4.3
libarchive-c==2.8
lief==0.9.0
line-profiler==3.1.0
llvmlite==0.31.0
locket==0.2.0
lru-dict==1.1.6
lxml==4.4.2
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
memory-profiler==0.55.0
mistune==0.8.4
mkl-fft
mkl-random
mkl-service==2.3.0
mlflow==1.2.0
mock==3.0.5
more-itertools==8.0.2
mpi4py==3.0.3
mpmath==1.1.0
msgpack==0.6.1
multipledispatch==0.6.0
munch==2.5.0
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.4
nevergrad==0.4.2.post2
nlopt==2.7.0
nltk==3.4.5
nose==1.3.7
notebook==6.0.2
numba==0.47.0
numexpr==2.7.0
numpy==1.18.0
numpydoc==0.9.2
oauthlib==3.1.0
olefile==0.46
opencv-python==3.4.5.20
openpyxl==3.0.2
opt-einsum==3.1.0
osqp==0.6.1
packaging==20.0
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.2
partd==1.1.0
path==13.1.0
pathlib2==2.3.5
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
pipreqs==0.4.10
pkginfo==1.5.0.1
pluggy==0.13.1
ply==3.11
pretrainedmodels==0.7.4
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.9.0
proxmin==0.6.11
psutil==5.6.7
ptflops==0.6.1
ptyprocess==0.6.0
py==1.8.1
py3nvml==0.2.6
pyasn1==0.4.8
pyasn1-modules==0.2.7
pyboy==1.0.1
pycocotools==2.0.6
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycryptodome==3.9.9
pydocstyle==4.0.1
pyflakes==2.1.1
pyglet==1.5.0
Pygments==2.5.2
pylint==2.4.4
pylru==1.2.0
pyodbc==4.0.27
pyOpenSSL==19.1.0
pyOpt==0.84
pyparsing==2.4.6
pyrsistent==0.15.6
PySDL2==0.9.7
PySocks==1.7.1
pytest==5.3.2
pytest-arraydiff==0.3
pytest-astropy==0.7.0
pytest-astropy-header==0.1.1
pytest-doctestplus==0.5.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-dateutil==2.8.1
python-editor==1.0.4
python-gflags==3.1.2
python-jsonrpc-server==0.3.2
python-language-server==0.31.2
pytz==2019.3
PyWavelets==1.1.1
pyxdg==0.26
PyYAML==5.2
pyzmq==18.1.0
QDarkStyle==2.7
QtAwesome==0.6.0
qtconsole==4.6.0
QtPy==1.9.0
querystring-parser==1.2.4
randomgen==1.15.0
requests==2.25.1
requests-oauthlib==1.3.0
robustness==1.2.1.post2
rope==0.14.0
rsa==4.0
Rtree==0.8.3
ruamel-yaml==0.15.87
scikit-image==0.15.0
scikit-learn==0.22.1
scipy==1.5.4
scs==2.1.2
seaborn==0.9.0
SecretStorage==3.1.1
Send2Trash==1.5.0
shutilwhich==1.1.0
simplegeneric==0.8.1
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.13.0
smmap==3.0.1
smmap2==2.0.5
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.9.5
Sphinx==2.3.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
spyder==4.0.1
spyder-kernels==1.8.1
SQLAlchemy==1.3.12
sqlparse==0.3.0
statsmodels==0.10.1
sympy==1.5.1
tables==3.6.1
tabulate==0.8.3
tblib==1.6.0
tempdir==0.7.1
tensorboard==1.12.2
tensorboardX==1.5
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
thop==0.0.31.post2004101309
toolz==0.10.0
# torch==1.7.1+cu110
#torchaudio==0.7.2
#torchfun==1.0.82
#torchstat==0.0.7
#torchvision==0.8.2+cu110
tornado==6.0.3
tqdm==4.41.1
traitlets==4.3.3
typing-extensions==3.7.4.1
ujson==1.35
unicodecsv==0.14.1
urllib3==1.25.7
watchdog==0.9.0
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
wurlitzer==2.0.0
xlrd==1.2.0
XlsxWriter==1.2.7
xlwt==1.3.0
xmltodict==0.12.0
yapf==0.28.0
yarg==0.1.9
zict==1.0.0
zipp==0.6.0
